# swagger.yaml
swagger: '2.0'
info:
  description: news のAPI仕様書
  version: 1.0.0
  title: News API
  contact:
    email: c8112002@gmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: 'localhost:8080'
basePath: /v1
tags:
  - name: articles
    description: 記事
  - name: tag_articles
    description: 指定されたタグの記事
  - name: add
    description: pocketに記事を追加
schemes:
  - http
paths:
  /articles:
    get:
      tags:
        - articles
      summary: Fetch articles
      description: ''
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: limit
          in: query
          description: 取得件数
          required: false
          type: number
          default: 5
        - name: offset
          in: query
          description: オフセット
          required: false
          type: number
          default: 0
      responses:
        '200':
          description: success
          schema:
            $ref: '#/definitions/Articles'
  '/articles/tag/{tag_name}':
    get:
      tags:
        - tag_articles
      summary: Fetch articles which have the specified tag
      description: ''
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: tag_name
          in: path
          description: タグ名
          required: true
          type: string
        - name: limit
          in: query
          description: 取得件数
          required: false
          type: number
          default: 5
        - name: offset
          in: query
          description: オフセット
          required: false
          type: number
          default: 0
      responses:
        '200':
          description: success
          schema:
            $ref: '#/definitions/Articles'
  /add:
    post:
      tags:
        - add
      summary: add an airticle to Pocket
      description: Pocketに記事を追加する
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: article_info
          in: body
          description: 記事情報
          required: true
          schema:
            $ref: '#/definitions/AddArticle'
      responses:
        '200':
          description: Article added.
          schema:
            $ref: '#/definitions/Article'
        '400':
          description: Bad requiest.
          schema:
            $ref: '#/definitions/Error'
        '422':
          description: Not Acceptable.
          schema:
            $ref: '#/definitions/NotAcceptableEntityError'
  /tags:
    get:
      tags:
        - tags
      summary: Fetch tags
      description: ''
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: limit
          in: query
          description: 取得件数
          required: false
          type: number
          default: 5
        - name: offset
          in: query
          description: オフセット
          required: false
          type: number
          default: 0
      responses:
        '200':
          description: success
          schema:
            $ref: '#/definitions/Tags'
definitions:
  Article:
    type: object
    properties:
      id:
        type: integer
      title:
        type: string
      url:
        type: string
      thumbnail:
        type: string
      published_at:
        type: string
        format: dateTime
      tags:
        type: array
        items:
          $ref: '#/definitions/Tag'
  Articles:
    type: object
    properties:
      articles:
        type: array
        items:
          $ref: '#/definitions/Article'
      isNext:
        type: boolean
      nextOffset:
        type: integer
      limit:
        type: integer
        description: リクエストに設定されたlimit
      offset:
        type: integer
        description: リクエストに設定されたoffset
  Tag:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
  Tags:
    type: object
    properties:
      tags:
        type: array
        items:
          $ref: '#/definitions/Tag'
      isNext:
        type: boolean
      nextOffset:
        type: integer
      limit:
        type: integer
        description: リクエストに設定されたlimit
      offset:
        type: integer
        description: リクエストに設定されたoffset
  AddArticle:
    type: object
    properties:
      article_id:
        type: integer
      tag_names:
        type: array
        description: 記事に紐付けるタグ名の配列
        items:
          type: string
  Error:
    type: object
    properties:
      code:
        type: integer
      message:
        type: string
  NotAcceptableEntityError:
    type: object
    properties:
      code:
        type: integer
      message:
        type: string
      errors:
        type: array
        items:
          type: object
          properties:
            code:
              type: integer
            parameter:
              type: string
              description: 不正なデータを渡したリクエストパラメータ名
            message:
              type: string